refactor atri-core
