+++ python-packages/atri-core/src/atri_core/__init__.py
@@ -1,2 +1,4 @@
-from .AtriComponent import AtriComponent
-from .AtriStyles import AtriStyles
+from .atri_component import AtriComponent
+from .atri_styles import AtriStyles
+
+__all__ = ["AtriComponent", "AtriStyles"]
